Fix mouse move and click with multiple monitors
Point y for `CGEventCreateMouseEvent` should be calculated by
the primary screen height.
`mainScreen` is the window with the keyboard focus.
The primary screen is the first object (index 0) in the array
returned by the screens.
Mouse location calculation was not valid with multiple monitors.
When mouse location is out of any screens, we move the location to
nearest screen edge.